+++ CHANGELOG.md
@@ -9,7 +9,9 @@
   Attention: Many units are changed to be more consistent with typical meteorological units. We now use `°C` for 
   temperatures. Also, length units are now separated in `length_short`, `length_medium` and `length_long` to get more
   reasonable decimals. Fore more information, see the new units chapter (usage/units) in the documentation.
-  
+
+### Fix
+- Filter out invalid underscore prefixed files
 
 ## 0.99.0 - 2024-12-30